Move filesystems to own file
