+++ src/components/mod.rs
@@ -1,3 +1,4 @@
 pub mod banner;
+pub mod filesystem;
 pub mod ssl_certs;
 pub mod uptime;
